Text renderer improvements.

+++ src/pokedex_screen.c
@@ -1934,7 +1934,7 @@ static int DexScreen_InputHandler_GetShoulderInput(void)
             return 2;
         else
             return 0;
-    case OPTIONS_BUTTON_MODE_HELP:
+    case (u8) OPTIONS_BUTTON_MODE_HELP:
     default:
         return 0;
     }
